requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
SQLAlchemy==2.0.43
psycopg2==2.9.10
python-dotenv==1.1.1
passlib==1.7.4
bcrypt==3.2.2
PyJWT==2.10.1
python-jose==3.5.0
python-multipart==0.0.20
pydantic==2.11.9
pydantic-settings==2.10.1
anyio==4.11.0
typing_extensions==4.15.0
click==8.3.0
cffi==2.0.0
cryptography==46.0.1
six==1.17.0
h11==0.16.0
sniffio==1.3.1
